# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cryptex Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password (authenticated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── User ──────────────────────────────────────────────────────────────────
...
  /exchange/orders:
    get:
      operationId: listExchangeOrders
      tags: [exchange]
      summary: List user exchange orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Exchange orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeOrderList"
    post:
      operationId: createExchangeOrder
      tags: [exchange]
      summary: Create exchange order (sell USDT for EUR)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeOrder"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /exchange/orders/{id}:
    get:
      operationId: getExchangeOrder
      tags: [exchange]
      summary: Get exchange order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exchange order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Savings ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth Schemas ─────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
      required: [email, password, confirmPassword]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
      required: [email, password]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
      required: [token, password]
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [currentPassword, newPassword]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        isAdmin:
          type: boolean
        availableEur:
          type: number
        lockedEur:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, email, isAdmin, availableEur, lockedEur, createdAt]
...
    ExchangeOrderInput:
      type: object
      properties:
        usdtAmount:
          type: number
          minimum: 50
          maximum: 50000
        txid:
          type: string
      required: [usdtAmount, txid]
    ExchangeOrder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        usdtAmount:
          type: number
        eurAmount:
          type: number
        txid:
          type: string
        status:
          type: string
          enum: [pending, completed, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, userId, usdtAmount, eurAmount, txid, status, createdAt]
    ExchangeOrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/ExchangeOrder"
        total:
          type: integer
      required: [orders, total]

    # ── Savings Schemas ───────────────────────────────────────────────────────
...